+++ NCU_BookShop/src/main/java/servlet/AdminGoodsEditServlet.java
@@ -42,7 +42,7 @@ public class AdminGoodsEditServlet extends HttpServlet {
                             g.setName(item.getString("utf-8"));
                             break;
                         case "price":
-                            g.setPrice(Float.parseFloat(item.getString("utf-8")));
+                            g.setPrice(Integer.parseInt(item.getString("utf-8")));
                             break;
                         case "intro":
                             g.setIntro(item.getString("utf-8"));
